Modified oeLogin
Modification done on oeLogin in icrashSystemImpl (server) to match with
new spec. Also added a line in DtString, but commented it. Not sure if
we are allowed to change the Dts. If you uncomment my line, then comment
everything else, since my modification makes everything else obsolete
while still making the same action.